Day one of trying to finish this thing. Added analyzer class. Added FFT algorithm. Added compilation targets for various platforms. Added global variables. Now using 16-bit integers for processing instead of floats (prepping for Teensy port).

+++ src/DelayLine.cpp
@@ -28,8 +28,8 @@ void DelayLine::tick() {
   }
 }
 
-float DelayLine::process() {
-  float sample;
+int16_t DelayLine::process() {
+  int16_t sample;
   // Run the delay line
   sample = x->getSample(position - delayTime) + ((y->getSample(position - delayTime - 1) + y->getSample(position - delayTime)) * 0.5 * feedback);
